Renamed js_passes
